+++ src/components/RegistrationForm/sub-components/Finish.tsx
