+++ src/components/RegistrationForm/sub-components/index.ts
